requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.94.0
anyio==4.12.1
async-timeout==4.0.3
attrs==26.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.28
langchain-anthropic==0.3.22
langchain-community==0.3.31
langchain-core==0.3.84
langchain-text-splitters==0.3.11
langsmith==0.4.37
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.19.0
networkx==3.2.1
numpy==2.0.2
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pgvector==0.4.2
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.6
psycopg2-binary==2.9.11
pyarrow==21.0.0
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydeck==0.9.1
pypdf==6.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.49
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.8.0
tornado==6.5.5
tqdm==4.67.3
transformers==4.57.6
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
yarl==1.22.0
zstandard==0.25.0
